API work
This commit adds a lot of the GraphQL mutations and types that will
eventually make up the overall API of the application.

These were generated by the graphql gem for Ruby and will need some
additional massaging to get things working as intended.

One thing to keep in mind is that the gem creates many `_id` fields as
Integers rather than IDs and the frontend GQL queries are very specific
about not massaging an ID (which is functionally a String!) into an
integer.

That is to say we'll need to manually change `_id` fields on types,
mutations, and probably input_types to be a GQL `ID` rather than
`Integer`.