Formalize the New/Edit Unit Components
This commit settles on how the New/Edit form will work and look.
Currently we'll be using a shared UnitFormFields component to handle the
form fields that the NewUnit and EditUnit components will display along
with their own custom handlers for the submit logic.

TODO: Need to set up the EditUnit to query the unit on page load.