Cleanup of Unit Components and State
In order to better handle Units this commit makes the Unit index
component act as a manager for the currentUnit state - setting the unit
when at a location that provides a unitId and clearing it whenever the
Unit components unmount.

Additionally this commit cleans up the UnitFormFields to work and behave
appropriately for both the New and Edit components.