Convert to data router in React Router, automate Breadcrumbs
In order to keep up with the way react router is heading, this commit
makes sure that we are using a full data router so that we have access
to attributes like loader for more easily handling loading data on a
route.

Additionally we automate (to an extent) the creation of breadcrumbs by
making it so that you can add a hash to the route handle prop so that
you can pass in a name that will display as the breadcrumb name for that
area. Dynamic names will need to pass in some kind of state to the
Breadcrumbs and will need to be added to the dynamicNames hash in
breadcrumbs.jsx to work properly.