Payment Logic
In order to allow for the basic creation of new payments this commit
adds the API logic for accepting a Graphql PaymentCreateInput and
creating the nested RentalAgreementPayments.

Also begins some work on refactoring some older pieces so that their
state is less intertwined when feasible.

Ultimately instead of doing things like passing in a "values" props
with everything would like to instead pass in simple props like a
"value" prop and a "onChange" prop that could be destructured in the
receiving component and used to avoid having to have knowledge of the
state of the parent component.