Minimum Viable Product
In order to allow the recording of owned properties, renters of those
properties, and payments made by those renters this commit represents
the completion of a minimum viable product.

There are a handful of basic todos that need for some simple fit and
finish and then some larger things that coudld be done in the future
such as listing out customer rentals or payments on the Customers show
page.