Table Changes and UI Cleanup
In order to allow users to sort table results this commit adds in the
EnhancedTable component which can be used to make relatively quick and
easy sortable tables.

Additionally the UI was generally cleaned up as well as the removal of
unused files and gems.

It also contains some of the initial setup for what will eventually be a
way to generate invoices for customers.