Change Next Due Date on Invoice Changes
In order to automate the process of changing next due dates whenever
invoices are updated or created this commit makes it so that when an
invoice_item is created it will automatically push the next_due_date of
the rental agreement.

In the event an invoice is updated it will modify the next_due_date
based on the difference of the new and old item_count.