Navigation Cleanup, Moved check number field, Preserve table ordering
In order to make the navigation cleaner this commit makes it (at least
on Payments) that when you complete a payment or cancel it you are
navigated either back a page OR navigated to the root for Payments if
there is no history.

Additionally the Check Number field was moved to be before the amount
field as this is apparently more in line with how the data entry would
normally go.

Finally the table ordering for the EnhancedTable sections is now
preserved via cookies.