Reset Invoice State on Index
Whenever a user goes from an Invoice to the index list, to a different
invoice it was preserving some of the previous invoice items across the
state. This commit makes it so that we nullify the invoice when we go to
the index.

Additionally this commit fixes an error where we were using merge rather
than assign on initialize. This could cause issues where states would
begin to overlap rather than truly initialize/reset a value. E.g. as
with the problem above we were getting objects in our invoice_items
array we didn't expect due to them being included as merge kept them
rather than overwriting them when updating the state.